added color scheme initializer
